创建纹理并使用。通过DefaultPool创建的纹理是无法unlockable的，需要一个SystemmemPool创建的临时纹理对象来更新纹理。

+++ Win32TestBed/particle.cpp
@@ -26,7 +26,7 @@ extern float RandomRangeF(float  min, float max);
 
 void ParticleInstance::SpawnParticle()
 {
-	if (m_particles.size() <= m_particleCount)
+	if (m_particles.size() <= (unsigned int)m_particleCount)
 	{
 		m_particles.push_back(Particle());
 	}
